Upgrade auto scroll feature
- On looking at the chat view from another view, chat box will be
automatically scrolled to the bottom.
- If a user is already scrolled to the bottom and a new message comes
in, the chat box will display the new message and scroll to the bottom
again so that the new message will be visible.
- If a user has scrolled up, and a new message comes in, the chat box
will NOT auto scroll, because that would be really obnoxious for the
user.

consulted with @marcpelve on integration with socket.io code and
handlers